Bugfix
The Spieler.class has always changed due to the fact, that it needs to be changed for programming.

The Problem lied within the Serialization. The Level-File serialized the Character-Object, which was giving an Error.

I instead changed the Serialization to the Player.class, which shouldn't change too much in the future.

TODO: Re-Export all the Levels
TODO: Make the next Chapters accessible after Level 7.
TODO: Make the Campaignlevels count again.